requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-ollama>=0.1.0
langchain-google-genai>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
